+++ sdk/networkcloud/Azure.ResourceManager.NetworkCloud/tests/Scenario/NetworkCloudClustersTests.cs
@@ -74,6 +74,22 @@ namespace Azure.ResourceManager.NetworkCloud.Tests.ScenarioTests
                         },
                     },
                 },
+                AnalyticsOutputSettings = new AnalyticsOutputSettings
+                {
+                    AnalyticsWorkspaceId = new ResourceIdentifier("/subscriptions/a3eeb848-665a-4dbf-80a4-eb460930fb23/resourceGroups/sdk-testing_id-simdev-3153212/providers/Microsoft.OperationalInsights/workspaces/simdev-3153212-law"),
+                    AssociatedIdentity = new ManagedServiceIdentitySelector
+                    {
+                        UserAssignedIdentityResourceId = new ResourceIdentifier("/subscriptions/a3eeb848-665a-4dbf-80a4-eb460930fb23/resourceGroups/sdk-testing_id-simdev-3153212/providers/Microsoft.ManagedIdentity/userAssignedIdentities/simdev-3153212-cluster-1-identity")
+                    }
+                },
+                CommandOutputSettings = new CommandOutputSettings
+                {
+                    AssociatedIdentity = new ManagedServiceIdentitySelector
+                    {
+                        UserAssignedIdentityResourceId = new ResourceIdentifier("/subscriptions/a3eeb848-665a-4dbf-80a4-eb460930fb23/resourceGroups/sdk-testing_id-simdev-3153212/providers/Microsoft.ManagedIdentity/userAssignedIdentities/simdev-3153212-cluster-1-identity")
+                    },
+                    ContainerUri = new Uri("https://myaccount.blob.core.windows.net/mycontainer?restype=container"),
+                },
                 Tags =
                 {
                     ["key1"] = "myvalue1",
@@ -133,6 +149,33 @@ namespace Azure.ResourceManager.NetworkCloud.Tests.ScenarioTests
             };
             var strategyResult = await clusterResource.UpdateAsync(WaitUntil.Completed, patch2);
             Assert.IsNotNull(strategyResult.Value);
+
+            // Patch Secret Archive Settings
+            NetworkCloudClusterPatch patch3 = new NetworkCloudClusterPatch()
+            {
+                SecretArchiveSettings = new SecretArchiveSettings
+                {
+                    VaultUri = new Uri("https://myaccount.blob.core.windows.net/mycontainer?restype=container"),
+                }
+            };
+            var secretArchiveResult = await clusterResource.UpdateAsync(WaitUntil.Completed, patch3);
+            Assert.IsNotNull(secretArchiveResult.Value);
+
+            // Patch VulnerabilityScanningContainerScan
+            try
+            {
+                NetworkCloudClusterPatch patch4 = new NetworkCloudClusterPatch()
+                {
+                    VulnerabilityScanningContainerScan = VulnerabilityScanningSettingsContainerScan.Enabled
+                };
+                var vulnerabilityScanResult = await clusterResource.UpdateAsync(WaitUntil.Completed, patch4);
+                Assert.IsNotNull(vulnerabilityScanResult.Value);
+            }
+            catch (Exception ex)
+            {
+                StringAssert.Contains("cluster conditions do not pass validation for cluster", ex.Message);
+                StringAssert.Contains("ClusterDeployedCondition is not True", ex.Message);
+            }
 
             // Cluster Update Version
             try