+++ sdk/networkcloud/Azure.ResourceManager.NetworkCloud/CHANGELOG.md
@@ -13,7 +13,10 @@
 ## 1.1.0 (2025-01-31)
 
 ### Features Added
-
+- This SDK version supports 2024-10-01-preview API version.
+- Cluster resource supports SecretArchiveSettings.
+- Cluster resource supports ClustersPatchVulnerabilityScanning with the ability to scan for cluster vulnerabilities.
+- 
 - Upgraded api-version tag from 'package-2023-07-01' to 'package-2024-07-01'. Tag detail available at https://github.com/Azure/azure-rest-api-specs/blob/f999652ecea2a4bddc2b08a113617e23e98f10d4/specification/networkcloud/resource-manager/readme.md.
 
 ### Other Changes